+++ src/parser/enrichers/trait/generic/PowerfulBuild.mjs
@@ -11,7 +11,6 @@ export default class PowerfulBuild extends DDBEnricherData {
         },
         changes: [
           DDBEnricherData.ChangeHelper.overrideChange("true", 20, "flags.dnd5e.powerfulBuild"),
-          // system.attributes.encumbrance.multipliers.overall
         ],
       },
     ];